Update to Bio-Formats 4.1.1.

+++ ie.tcd.imm.hits.image/trunk/ie.tcd.imm.hits.image/src/ie/tcd/imm/hits/image/loci/convert/ConvertToImageNodeModel.java
@@ -239,8 +239,8 @@ public class ConvertToImageNodeModel extends NodeModel {
 					final IMetadata metadata = pair.getLeft().getRight();
 					final double zDoubleValue = ((DoubleValue) zDataCell)
 							.getDoubleValue();
-					final float timeDiff = metadata.getPlaneTimingDeltaT(serie,
-							0, timeIndex)
+					final double timeDiff = metadata.getPlaneTimingDeltaT(
+							serie, 0, timeIndex)
 							- metadata.getPlaneTimingDeltaT(serie, 0, 0);
 					values.set(resultTimeIndex, new DoubleCell(timeDiff));
 					final FormatReader formatReader = pair.getLeft().getLeft();
